Drafted out the backend, ran into problems
You are in a maze of twisty little time/date formats, all
different.

I DO want to preserve the exact ntp timestamp rather than
translate to postgres's type(s).

That leads to a world of hurt. I need to stop and go look
at some other implementations - such as the round robin database
to see a good way to go.

Also got bit by featureitus - I'd like to use a geospacial
database - my vision is to be able to present bloat as
"height above ground" on a rotating globe with the time series(s).

My imagination shows the terminator line passing and behind it,
the bloated buffers gradulally becoming smaller and less noticable.